Moving RBT version of OSC to attic, renaming MBP version of OSC
