+++ examples/Cassie/BUILD.bazel
@@ -296,9 +296,9 @@ cc_binary(
         ":cassie_urdf",
         ":cassie_utils",
         "//attic/multibody:utils",
+        "//attic/systems/controllers:operational_space_control",
         "//examples/Cassie/osc",
         "//systems:robot_lcm_systems",
-        "//systems/controllers/osc:operational_space_control",
         "//systems/framework:lcm_driven_loop",
         "//systems/primitives",
         "@drake//:drake_shared_library",
