+++ systems/controllers/osc/operational_space_control.h
@@ -1,50 +1,52 @@
 #pragma once
 
+#include <memory>
+#include <string>
 #include <utility>
 #include <vector>
-#include <string>
-#include <memory>
+#include <drake/multibody/plant/multibody_plant.h>
+#include "drake/common/trajectories/exponential_plus_piecewise_polynomial.h"
+#include "drake/common/trajectories/piecewise_polynomial.h"
 #include "drake/systems/framework/diagram.h"
 #include "drake/systems/framework/diagram_builder.h"
 #include "drake/systems/framework/leaf_system.h"
-#include "drake/common/trajectories/piecewise_polynomial.h"
-#include "drake/common/trajectories/exponential_plus_piecewise_polynomial.h"
 #include "dairlib/lcmt_osc_output.hpp"
+
 #include "drake/solvers/mathematical_program.h"
 #include "drake/solvers/solve.h"
 
-#include "systems/framework/output_vector.h"
-#include "systems/controllers/osc/osc_tracking_data.h"
+#include "multibody/multibody_distance_constraint.h"
 #include "systems/controllers/control_utils.h"
+#include "systems/controllers/osc/osc_tracking_data_mbp.h"
+#include "systems/framework/output_vector.h"
 
-namespace dairlib {
-namespace systems {
-
-namespace controllers {
+namespace dairlib::systems::controllers {
 
 /// `OperationalSpaceControl` takes in desired trajectory in world frame and
 /// outputs torque command of the motors.
 
 /// Inputs of the constructor:
-///  - `tree_w_spr` a RigidBodyTree with springs
-///  - `tree_wo_spr` a RigidBodyTree without springs
+///  - `plant_w_spr` a MultibodyPlant with springs. If the full model of the
+///    plant does not have spring, then plant_w_spr and plant_wo_spr should
+///    refer to the same plant.
+///  - `plant_wo_spr` a MultibodyPlant without springs
 ///  - `used_with_finite_state_machine` a flag indicating whehter using osc with
 ///    finite state machine or not
 /// The springs here refer to the compliant components in the robots.
 
-/// OSC calculates feedback positions/velocities from `tree_w_spr`,
-/// but in the optimization it uses `tree_wo_spr`. The reason of using
-/// RigidBodyTree without spring is that the OSC cannot track desired
+/// OSC calculates feedback positions/velocities from `plant_w_spr`,
+/// but in the optimization it uses `plant_wo_spr`. The reason of using
+/// MultibodyPlant without spring is that the OSC cannot track desired
 /// acceleration instantaneously when springs exist. (relative degrees of 4)
 
 /// Requirement:
-///  - the joints name (except for the spring joints) in `tree_w_spr` must be
-///    the same as those of `tree_wo_spr`
-///  - the bodies in both RBT's should be the same. (to get jacobian from
-///    both trees)
+///  - the joints name (except for the spring joints) in `plant_w_spr` must be
+///    the same as those of `plant_wo_spr`
+///  - the bodies in both MBP's should be the same. (to get Jacobian from
+///    both plants)
 
-/// If the robot doesn't have any spring, the user can just pass two identical
-/// RigidBodyTree into the constructor.
+/// If the robot doesn't have any springs, the user can just pass two identical
+/// MultibodyPlants into the constructor.
 
 /// Users define
 ///     costs,
@@ -54,6 +56,7 @@ namespace controllers {
 
 /// Before adding desired trajectories to `OperationalSpaceControl` with the
 /// method `AddTrackingData`, users have to create
+///     `CenterOfMassTrackingData`,
 ///     `TransTaskSpaceTrackingData`,
 ///     `RotTaskSpaceTrackingData`,
 ///     and/or `JointSpaceTrackingData`.
@@ -67,7 +70,7 @@ namespace controllers {
 /// the outputs of trajectory blocks need to be of the derived classes of
 ///     drake::trajectories::Trajectory<double>
 /// such as `PiecewisePolynomial` and `ExponentialPlusPiecewisePolynomial`.
-/// The users can connect the output ports of the desired trajecotry blocks to
+/// The users can connect the output ports of the desired trajectory blocks to
 /// the corresponding input ports of `OperationalSpaceControl` by using
 /// the method get_tracking_data_input_port().
 
@@ -81,10 +84,18 @@ namespace controllers {
 
 class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
  public:
-  OperationalSpaceControl(const RigidBodyTree<double>& tree_w_spr,
-                          const RigidBodyTree<double>& tree_wo_spr,
-                          bool used_with_finite_state_machine = true,
-                          bool print_tracking_info = false);
+  OperationalSpaceControl(
+      const drake::multibody::MultibodyPlant<double>& plant_w_spr,
+      const drake::multibody::MultibodyPlant<double>& plant_wo_spr,
+      bool used_with_finite_state_machine = true,
+      bool print_tracking_info = false);
+
+  const drake::systems::OutputPort<double>& get_osc_output_port() const {
+    return this->get_output_port(osc_output_port_);
+  }
+  const drake::systems::OutputPort<double>& get_osc_debug_port() const {
+    return this->get_output_port(osc_debug_port_);
+  }
 
   // Input/output ports
   const drake::systems::InputPort<double>& get_robot_output_input_port() const {
@@ -94,48 +105,47 @@ class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
     return this->get_input_port(fsm_port_);
   }
   const drake::systems::InputPort<double>& get_tracking_data_input_port(
-      std::string name) const {
+      const std::string& name) const {
     return this->get_input_port(traj_name_to_port_index_map_.at(name));
-  }
-  const drake::systems::OutputPort<double>& get_osc_output_port() const {
-    return this->get_output_port(osc_output_port_);
-  }
-  const drake::systems::OutputPort<double>& get_osc_debug_port() const {
-    return this->get_output_port(osc_debug_port_);
   }
 
   // Cost methods
-  void SetInputCost(Eigen::MatrixXd W) {W_input_ = W;}
-  void SetAccelerationCostForAllJoints(Eigen::MatrixXd W) {W_joint_accel_ = W;}
-  void AddAccelerationCost(std::string joint_vel_name, double w);
+  void SetInputCost(const Eigen::MatrixXd& W) { W_input_ = W; }
+  void SetAccelerationCostForAllJoints(const Eigen::MatrixXd& W) {
+    W_joint_accel_ = W;
+  }
+  void AddAccelerationCost(const std::string& joint_vel_name, double w);
 
   // Constraint methods
-  void DisableAcutationConstraint() {with_input_constraints_ = false;}
-  void SetContactFriction(double mu) {mu_ = mu;}
+  void DisableAcutationConstraint() { with_input_constraints_ = false; }
+  void SetContactFriction(double mu) { mu_ = mu; }
   void SetWeightOfSoftContactConstraint(double w_soft_constraint) {
     w_soft_constraint_ = w_soft_constraint;
   }
-  void AddContactPoint(std::string body_name, Eigen::VectorXd pt_on_body);
-  void AddStateAndContactPoint(int state,
-                               std::string body_name,
+  void AddContactPoint(const std::string& body_name,
+                       const Eigen::VectorXd& pt_on_body);
+  void AddStateAndContactPoint(int state, std::string body_name,
                                Eigen::VectorXd pt_on_body);
-
+  void AddDistanceConstraint(
+      multibody::MultibodyDistanceConstraint& constraint);
   // Tracking data methods
   /// The third argument is used to set a period in which OSC does not track the
   /// desired traj (the period starts when the finite state machine switches to
   /// a new state)
-  void AddTrackingData(OscTrackingData* tracking_data,
-      double t_lb = 0, double t_ub = std::numeric_limits<double>::infinity());
-  void AddConstTrackingData(OscTrackingData* tracking_data, Eigen::VectorXd v,
-      double t_lb = 0, double t_ub = std::numeric_limits<double>::infinity());
-  std::vector<OscTrackingData*>* GetAllTrackingData(){
+  void AddTrackingData(OscTrackingDataMBP* tracking_data, double t_lb = 0,
+                       double t_ub = std::numeric_limits<double>::infinity());
+  void AddConstTrackingData(
+      OscTrackingDataMBP* tracking_data, const Eigen::VectorXd& v, double t_lb
+      = 0,
+      double t_ub = std::numeric_limits<double>::infinity());
+  std::vector<OscTrackingDataMBP*>* GetAllTrackingData() {
     return tracking_data_vec_.get();
   }
-  OscTrackingData* GetTrackingDataByIndex(int index) {
+  OscTrackingDataMBP* GetTrackingDataByIndex(int index) {
     return tracking_data_vec_->at(index);
   }
 
-  // Osc leafsystem builder
+  // OSC LeafSystem builder
   void Build();
 
  private:
@@ -144,27 +154,29 @@ class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
   void CheckConstraintSettings();
 
   // Get solution of OSC
-  Eigen::VectorXd SolveQp(Eigen::VectorXd x_w_spr, Eigen::VectorXd x_wo_spr,
-      const drake::systems::Context<double>& context, double t,
-      int fsm_state, double time_since_last_state_switch) const;
+  Eigen::VectorXd SolveQp(const Eigen::VectorXd& x_w_spr,
+                          const Eigen::VectorXd& x_wo_spr,
+                          const drake::systems::Context<double>& context,
+                          double t, int fsm_state,
+                          double time_since_last_state_switch) const;
 
   // Discrete update that stores the previous state transition time
   drake::systems::EventStatus DiscreteVariableUpdate(
       const drake::systems::Context<double>& context,
       drake::systems::DiscreteValues<double>* discrete_state) const;
 
+  void AssignOscLcmOutput(const drake::systems::Context<double>& context,
+                          dairlib::lcmt_osc_output* output) const;
+
   // Output function
   void CalcOptimalInput(const drake::systems::Context<double>& context,
                         systems::TimestampedVector<double>* control) const;
 
-  void AssignOscLcmOutput(const drake::systems::Context<double>& context,
-                          dairlib::lcmt_osc_output* output) const;
-
   // Input/Output ports
+  int osc_debug_port_;
+  int osc_output_port_;
   int state_port_;
   int fsm_port_;
-  int osc_output_port_;
-  int osc_debug_port_;
 
   // Discrete update
   int prev_fsm_state_idx_;
@@ -177,11 +189,19 @@ class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
   // Map from (non-const) trajectory names to input port indices
   std::map<std::string, int> traj_name_to_port_index_map_;
 
-  // RBT's.
-  const RigidBodyTree<double>& tree_w_spr_;
-  const RigidBodyTree<double>& tree_wo_spr_;
+  // MBP's.
+  const drake::multibody::MultibodyPlant<double>& plant_w_spr_;
+  const drake::multibody::MultibodyPlant<double>& plant_wo_spr_;
 
-  // Size of position, velocity and input of the RBT without spring
+  // World frames
+  const drake::multibody::BodyFrame<double>& world_w_spr_;
+  const drake::multibody::BodyFrame<double>& world_wo_spr_;
+
+  // MBP context's
+  std::unique_ptr<drake::systems::Context<double>> context_w_spr_;
+  std::unique_ptr<drake::systems::Context<double>> context_wo_spr_;
+
+  // Size of position, velocity and input of the MBP without spring
   int n_q_;
   int n_v_;
   int n_u_;
@@ -190,11 +210,14 @@ class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
   int n_h_;
   int n_c_;
 
+  // Manually specified holonomic constraints (only valid for plants_wo_springs)
+  std::vector<multibody::MultibodyDistanceConstraint*> distance_constraints_;
+
   // robot input limits
   Eigen::VectorXd u_min_;
   Eigen::VectorXd u_max_;
 
-  // flag indicating whehter using osc with finite state machine or not
+  // flag indicating whether using osc with finite state machine or not
   bool used_with_finite_state_machine_;
 
   // flag indicating whether to print the tracking related values or not
@@ -220,16 +243,16 @@ class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
 
   // OSC cost members
   /// Using u cost would push the robot away from the fixed point, so the user
-  /// could consider usnig acceleration cost instead.
-  Eigen::MatrixXd W_input_;  // Input cost weight
+  /// could consider using acceleration cost instead.
+  Eigen::MatrixXd W_input_;        // Input cost weight
   Eigen::MatrixXd W_joint_accel_;  // Joint acceleration cost weight
 
   // OSC constraint members
   bool with_input_constraints_ = true;
 
   // (flat ground) Contact constraints and friction cone constraints
-  std::vector<int> body_index_ = {};
-  std::vector<Eigen::VectorXd> pt_on_body_ = {};
+  std::vector<drake::multibody::BodyIndex> body_indices_ = {};
+  std::vector<Eigen::VectorXd> pts_on_body_ = {};
   double mu_ = -1;  // Friction coefficients
   double w_soft_constraint_ = -1;
 
@@ -245,8 +268,9 @@ class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
   std::vector<bool> CalcActiveContactIndices(int fsm_state) const;
 
   // OSC tracking data (stored as a pointer because of caching)
-  std::unique_ptr<std::vector<OscTrackingData*>> tracking_data_vec_ =
-      std::make_unique<std::vector<OscTrackingData*>>();
+  std::unique_ptr<std::vector<OscTrackingDataMBP*>> tracking_data_vec_ =
+      std::make_unique<std::vector<OscTrackingDataMBP*>>();
+
   // Fixed position of constant trajectories
   std::vector<Eigen::VectorXd> fixed_position_vec_;
 
@@ -257,7 +281,4 @@ class OperationalSpaceControl : public drake::systems::LeafSystem<double> {
   std::vector<double> t_e_vec_;
 };
 
-
-}  // namespace controllers
-}  // namespace systems
-}  // namespace dairlib
+}  // namespace dairlib::systems::controllers
